+++ contracts/interpreter/ops/core/OpLoopN.sol
@@ -27,14 +27,14 @@ library OpLoopN {
         StackPointer stackTop_
     ) internal view returns (StackPointer) {
         unchecked {
-            uint n_ = Operand.unwrap(operand_) >> 12;
-            uint inputs_ = Operand.unwrap(operand_) & MASK_4BIT;
-            uint outputs_ = (Operand.unwrap(operand_) >> 4) & MASK_4BIT;
+            uint256 n_ = Operand.unwrap(operand_) >> 12;
+            uint256 inputs_ = Operand.unwrap(operand_) & MASK_4BIT;
+            uint256 outputs_ = (Operand.unwrap(operand_) >> 4) & MASK_4BIT;
             require(inputs_ >= outputs_, "LOOP_N_INPUTS");
             Operand callOperand_ = Operand.wrap(
                 Operand.unwrap(operand_) & MASK_12BIT
             );
-            for (uint i_ = 0; i_ < n_; i_++) {
+            for (uint256 i_ = 0; i_ < n_; i_++) {
                 stackTop_ = OpCall.integrity(
                     integrityCheckState_,
                     callOperand_,
